fix:修改bug，修改mysql redis配置

+++ mall-portal/src/main/java/com/macro/mall/portal/service/UmsMemberReceiveAddressService.java
@@ -13,7 +13,7 @@ public interface UmsMemberReceiveAddressService {
     /**
      * 添加收货地址
      */
-    int add(UmsMemberReceiveAddress address);
+    long add(UmsMemberReceiveAddress address);
 
     /**
      * 删除收货地址
